+++ ECourseWeb/src/main/java/com/ecourse/pojo/ReceiptDetail.java
@@ -37,7 +37,7 @@ public class ReceiptDetail implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "quantity")
     private Integer quantity;
     // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
@@ -53,19 +53,19 @@ public class ReceiptDetail implements Serializable {
     public ReceiptDetail() {
     }
 
-    public ReceiptDetail(Integer id) {
+    public ReceiptDetail(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
